saya-bot/requirements.txt
openai==1.3.0
python-telegram-bot==20.7
speechrecognition==3.10.0
gtts==2.4.0
Pillow==10.1.0
requests==2.31.0
google-generativeai==0.3.2
pydub==0.25.1